chore: move donations to global data
